+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ---
 
+#### 21.02.2018
+
+* [Rust] Railways (Setup & tests) (Herring)
+
 #### 13.02.2018
 
 * [Rust] Bookshelf problem (Shelf capacity) (Reddit)